api/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.1
django-cors-headers==3.5.0
django-filter==2.3.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
pycodestyle==2.6.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.3.1
toml==0.10.1
